==> Data Preprocessing and SRW/03. Feature Extraction.ipynb <==
from collections import Counter
import networkx as nx
import pandas as pd
import pickle
import pprint

data = pd.read_csv('./raw/kubernetes.csv')

pr_id_list = data['pr_id'].unique()

### Get Number of Reviews For Pull Request:

def get_num_reviews(pr_id, data):
    
    reviewer_list={}
    pr_df = data.loc[data.pr_id == pr_id, :]
    
    commenter_list = list(pr_df.commenter_id.values)
    commenter_set = set(commenter_list)
    
    head_commiter_set = set(pr_df.head_commit_committer_id.values)
    head_author_set = set(pr_df.head_commit_author_id.values)
    pr_owner_set = (head_author_set | head_commiter_set) & commenter_set
    reviewer_set = commenter_set - pr_owner_set
    
    buf_list = []
    for c in commenter_list:
        if c not in pr_owner_set:
            buf_list += [c]
    
    commenter_freq = {}
    for reviewer in buf_list:
        if reviewer not in commenter_freq:
            commenter_freq.update({reviewer: 1})
        else:
            commenter_freq[reviewer] += 1

    return commenter_freq

for pr_id in pr_id_list[0:5]: 
    res = get_num_reviews(pr_id, data)
    print('PR_ID: {}, REVIEWERS: {}'.format(pr_id, res))

### Number of comments made by Reviewer r on Pull Request p:

def ncomments_on_pr_by_reviewer(pr_id, reviewer_id, data):
    
    res = get_num_reviews(pr_id, data)
    
    if reviewer_id in res.keys():
        return res[reviewer_id]
    else:
        return 0

for pr_id in pr_id_list[0:5]:
    for reviewer in get_num_reviews(pr_id, data).keys():
        res = ncomments_on_pr_by_reviewer(pr_id, reviewer, data)
        print('PR_ID: {}, REVIEWER: {}, NUM_COMMENTS: {}'.format(pr_id, reviewer, res))

### Pull-Requests submitted by Reviewer r to the repository:

def npullreq_on_repo_by_reviewer(reviewer, data):
    df = set(data[data.commenter_id == reviewer]['head_repo_id'])
    return len(df)

entities = None
with open('./cache/arr/entities.pkl', 'rb') as file:
    entities = pickle.load(file)
    
entity_metadata = None
with open('./cache/dicts/entity_metadata.pkl', 'rb') as file:
    entity_metadata = pickle.load(file)

for e in entities[0:10]:
    print('ENTITY: {}, NUM_PULL_REQ: {}'.format(e, npullreq_on_repo_by_reviewer(e, data)))

### Reviewers Count:

reviewer_count ={}
for pr_id in pr_id_list:
    reviewer_count[pr_id] = get_num_reviews(pr_id,data)

### List of Pull Requests Submitted by Reviewer r to the repository:

#### Load Entities:

entities = None
with open('./cache/arr/entities.pkl', 'rb') as file:
    entities = pickle.load(file)

#### Finding the List of Pull Requests:

total_pr_reviewed = {}

for reviewer in entities:
    for pr_id in set(pr_id_list):
        if pr_id in reviewer_count.keys():
            if reviewer in list(reviewer_count[pr_id].keys()):

                if reviewer not in total_pr_reviewed.keys():

                    total_pr_reviewed.update({
                        reviewer: {
                            'pr_list': [pr_id], 
                            'total_counts': 1
                        }
                    })

                else:

                    total_pr_reviewed[reviewer]['pr_list'] += [pr_id]
                    total_pr_reviewed[reviewer]['total_counts'] += 1

total_pr_count = {}

for (rev, pr) in total_pr_reviewed.items():
    total_pr_count[rev] = len(pr)

#### Reviewers List:

reviewers = list(total_pr_count.keys())

### Export Reviewers:

export_list = False

if export_list == True:
    with open('./cache/arr/reviewers.pkl', 'wb') as file:
        pickle.dump(reviewers, file)

### Total Reviewed Items by Reviewer:

for reviewer in reviewers[0:5]:
    
    res = total_pr_reviewed[reviewer]
    
    print('======================================================')
    print('REVIEWER_ID:', reviewer)
    print('REVIEWED_PR_LIST:', res['pr_list'])
    print('REVIEWED_PR_COUNT:', res['total_counts'])
